requirements.txt
Flask==2.3.2
PyMuPDF==1.22.5
google-cloud-translate==3.12.0